requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# HTTP client for async requests
httpx>=0.24.0

# Web scraping and content extraction
trafilatura>=2.0.0
readability-lxml>=0.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Audio processing
pydub>=0.25.0
gTTS>=2.3.0

# AI/LLM providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Web search
duckduckgo-search>=4.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 